+++ src/openApi/v2/parser/getOperationName.ts
@@ -6,6 +6,7 @@ import camelCase from 'camelcase';
  * on a generated name from the URL
  */
 export const getOperationName = (url: string, method: string, operationId?: string): string => {
+    console.log(url, method, operationId);
     if (operationId) {
         return camelCase(
             operationId
@@ -15,10 +16,18 @@ export const getOperationName = (url: string, method: string, operationId?: stri
         );
     }
 
-    const urlWithoutPlaceholders = url
+    let urlWithoutPlaceholders = url
         .replace(/[^/]*?{api-version}.*?\//g, '')
         .replace(/{(.*?)}/g, '')
         .replace(/\//g, '-');
+
+    // Remove invalid characters
+    urlWithoutPlaceholders = urlWithoutPlaceholders
+        .split('-')
+        .map(part => {
+            return part.replace(/^[^a-z]+/gi, '');
+        })
+        .join('-');
 
     return camelCase(`${method}-${urlWithoutPlaceholders}`);
 };