+++ src/openApi/v3/parser/getOperationName.ts
@@ -6,19 +6,29 @@ import camelCase from 'camelcase';
  * on a generated name from the URL
  */
 export const getOperationName = (url: string, method: string, operationId?: string): string => {
+    console.log(url, method, operationId);
+
     if (operationId) {
         return camelCase(
             operationId
-                .replace(/^[^a-zA-Z]+/g, '')
+                .replace(/^[^a-z]+/gi, '')
                 .replace(/[^\w\-]+/g, '-')
                 .trim()
         );
     }
 
-    const urlWithoutPlaceholders = url
+    let urlWithoutPlaceholders = url
         .replace(/[^/]*?{api-version}.*?\//g, '')
         .replace(/{(.*?)}/g, '')
         .replace(/\//g, '-');
+
+    // Remove invalid characters
+    urlWithoutPlaceholders = urlWithoutPlaceholders
+        .split('-')
+        .map(part => {
+            return part.replace(/^[^a-z]+/gi, '');
+        })
+        .join('-');
 
     return camelCase(`${method}-${urlWithoutPlaceholders}`);
 };