+++ src/openApi/v3/parser/getOperationParameters.ts
@@ -57,5 +57,6 @@ export const getOperationParameters = (openApi: OpenApi, parameters: OpenApiPara
             }
         }
     });
+
     return operationParameters;
 };